Fix Mono crash when getting the MSBuild version
The implementation of GetVersionInfo() in Mono doesn't set the
ProductVersion property, so ProjectCollection.Version crashes
when trying to construct the Version object from an empty string.

As a workaround, the Version object can be created using the
individual components of the product version, which are properly
filled by Mono. That's also more efficient that relying on
string version parsing.